requirements.txt
bitsandbytes==0.41.1
dadaptation==3.1
diffusers==0.31.0
flux==1.3.5
ipython==8.12.3
ipython==8.7.0
lion_pytorch==0.1.2
lpips==0.1.4
matplotlib==3.6.2
numpy==1.23.5
opencv_python==4.5.5.64
opencv_python_headless==4.7.0.68
Pillow==10.1.0
Pillow==11.2.1
prodigyopt==1.0
pydantic==2.11.3
PyYAML==6.0.1
PyYAML==6.0.2
safetensors==0.5.3
torch==2.5.0+cu121
torch_xla==2.6.1
torchvision==0.20.0+cu121
tqdm==4.64.1
transformers==4.47.1
